Pre-1.3.2 Beta, Update 3 -------------------------------------------------------- -New MobTargetterTask - replacing MobTargettingTask - to work with the pathfinder in MCPath -Linked ZAMob and ZACharacter through ZALiving -Objects can now implement Targettable, allowing them to be set as a target for the targetter. This requires them to provide a location every runthrough of the targetter task -Removed VehicleExitEvent because it didn't work as intended -Pressing L-Shift in last stand no longer kicks you. In fact, L-Shift does nothing, as intended -Updated MCShot and MCPath, which may not be different version numbers but have been changed internally

+++ src/com/github/jamesnorris/ablockalypse/event/bukkit/EntityDeath.java
@@ -14,8 +14,8 @@ import org.bukkit.event.entity.EntityDeathEvent;
 
 import com.github.jamesnorris.ablockalypse.Ablockalypse;
 import com.github.jamesnorris.ablockalypse.DataContainer;
-import com.github.jamesnorris.ablockalypse.aspect.entity.ZAMob;
-import com.github.jamesnorris.ablockalypse.aspect.entity.ZAPlayer;
+import com.github.jamesnorris.ablockalypse.aspect.ZAMob;
+import com.github.jamesnorris.ablockalypse.aspect.ZAPlayer;
 import com.github.jamesnorris.ablockalypse.enumerated.PowerupType;
 import com.github.jamesnorris.ablockalypse.enumerated.Setting;
 
@@ -35,7 +35,7 @@ public class EntityDeath implements Listener {
             killer.setFoodLevel(20);
         }
         int chance = rand.nextInt(100) + 1;
-        if (chance <= (Integer) Setting.POWERUP_CHANCE.getSetting()) {
+        if (chance <= (Double) Setting.POWERUP_CHANCE.getSetting()) {
             zap.givePowerup(PowerupType.getById(rand.nextInt(5) + 1), killer);
         }
     }
